requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
faiss-cpu
python-dotenv
numpy
sqlalchemy
aiosqlite